Refactor JupyterBookHomepage -> JupyterBookPage
And move to astropylibrarian.reducers. This dataclass now becomes the
basis for reducing content in all JupyterBook pages, not just the
homepage.
